Extract big queries + discipline polling behaviour
